refactor: 긴급대응 메인 페이지 관련 컴포넌트들 components 폴더로 분리
