+++ src/pages/Emergency/EmergencyMain/components/EmergencyResponseStart.tsx
@@ -1,4 +1,4 @@
-import Button from "../../../components/Button/Button";
+import Button from "../../../../components/Button/Button";
 
 const EmergencyResponseStart = () => {
   return (
