+++ src/pages/Emergency/EmergencyMain/components/OrganCard.tsx
